Talk Rocking Agile
Do you believe that agile is only for IT teams? You’re wrong. I have
been using agile methodologies for a while with IT teams and as a
passioned amateur musician having played in several band i thought "How
the hell can we work as an agile team in a rock band?". I had the
opportunity 2 years ago to start a new band and try this from scratch.
I’ll show you the journey so far and how we survived the constant
changes and transformed into a better, motivated and more efficient
band adapting agile methodologies and procedures like TDD, standups,
retrospectives, etc.