Strikers Edge Post Mortem
The story and analysis of what went right and wrong with Strikers Edge: a multiplayer medieval "dodgebrawl" game. From it's humble gamejam origins, to winner of "Prémios PlayStation", to fully fledged PS4 and PC game. We'll talk about the hurdles the team went through, the technology, art, design, marketing and business decisions and it's impact, the successes and failures and all the lessons we learned in the process. An inside view on the experience of becoming an independent game developer.