Data Structures That Power Your Database
Data Structures That Power Your Database
========================================

* Speaker   : Pedro Tavares
* Available : Anytime
* Length    : 45 minutes
* Language  : English

Description
-----------

On the most fundamental level, a database needs to do two things: when
you give it some data, it should store the data, and when you ask it
again later, it should give the data back to you. Why should you, as an
application developer, care how the database handles storage and
retrieval internally? In order to select the storage engine that is
appropriate for your application, you need to have a rough idea of what
the storage engine is doing under the hood. In this talk, we’ll discuss
and examine some core data structures such as Hash Indexes, SSTables,
LSM-Trees, and B-Trees, that are used in the traditional relational
databases and NoSQL databases.

Speaker Bio
-----------

Pedro is a passionate Software Engineer currently working at Talkdesk.
Distributed Systems and Software Engineering best practices bring a big
smile on his face. He likes to read books and scientific papers on
Computer Science and he's currently running the Porto's chapter of
Papers We Love.

Links
----------

* Blog: https://ordepdev.me
* GitHub: https://github.com/ordepdev
* Twitter: https://twitter.com/ordepdev