Struggling with AWS
Struggling with AWS
=========================

* Speaker   : **Gabriel Poca**
* Available : **second day, 9am-7pm**
* Length    : **30 minutes**
* Language  : **English**

Description
-----------

In this talk I will drop every bit of knowledge I wish I had when my team started deploying to AWS. This is not an introduction, nor a love story, I'm going deep into how AWS works. I'm here to talk about what did work and what didn't when it comes to running services on the cloud, setting up a continuous delivery service, logging, error tracking, debugging live environments, secrets management, policies, preparing for failure and monitoring the whole thing.

The information in this talk will be based on real life experience deploying Elixir and JavaScript, running blockchain nodes and securing blockchain private keys in the cloud. With this, I hope to shed some light in the daunting path to deploying to AWS.

Speaker Bio
-----------

I don't have a fancy title to show of. I'm a developer and my work goes from hiring to cloud infrastructure. For the past year I've been focused on building the infrastructure for UTRUST (https://utrust.com/), which is a payment gateway for blockchains. I also write a lot of Elixir and JavaScript. In the past I wrote a lot of Ruby as well.

Links
-----

* Blog: https://gabrielpoca.com/
* Company: https://subvisual.co/
* GitHub: https://github.com/gabrielpoca