Beyond Code: how the right side of the brain can guide your journey to be a better Engineer
Being a solid SW Engineer is much more than just writing good code, knowing the latest languages or be aware of the hottest trends. Your career goals and the companies you work for also shape your future as they define much of the path you'll thread. There's always something to be learned with everyone that crosses your path with or in your current role. Sometimes these lessons aren't easy on us or the ones we'd like to learn but there's certainly virtue in accepting with grace and gratitude.