PWAs and Fighting Against App Stores
- Speaker : Gabriel Poça
- Available : first day, second day
- Length : 30 minutes
- Language : English

This talk is about what we can do with the open web. There are two parts to it.

The first part is smaller and philosophical. I'll talk about the open web and
its importance. My purpose is to show how broken the mobile world is, and how
dangerous the play store and app store are. How we've been in this situation
before, and what we should be doing to avoid getting into it again. These
stores are not open platforms, but most people cannot choose not
to use them. There have been plenty of antitrust lawsuits filed against Apple
and Google but nothing seems to change. Shouldn't we stop feeding the giant
before it gets too big?

The second part of this talk is more technical. I'll show what's available on
the web right now, and why it's probably enough for most of your apps. We'll
talk about deploys, notifications, offline support, media access, file system
access, etc. There's a lot in there. We will go over the good and the bad,
clarifying the limitations of each technology. I'll talk about what's to come
and to which platform. I also show what you can do right now to leverage both
web and native technology while you wait for support on the web.

The information in this talk is based on real life experience building PWAs for
many years now.

I don't have a fancy title to show of. I'm a developer and my work goes from
hiring to cloud infrastructure. For the past two years I've been focused on
building the infrastructure for UTRUST (https://utrust.com/), which is a
payment gateway for blockchains. I write a lot of Elixir and JavaScript. I care
about the social impact software has, and how the choices we make today impact
other people and the next generation.

- Blog: https://gabrielpoca.com/
- Company: https://subvisual.co/
- GitHub: https://github.com/gabrielpoca