Hooking up with React Hooks =================================================
* Speaker   : **João Dias** and **Mickael Costa**
* Available : **All days**
* Length    : **90 minutes**
* Language  : **English**

Description
-----------

In the last months, React Hooks changed how we write and think about reusable and straightforward code. This training workshop will cover the fundamentals of React Hooks, how to convert your codebase and also how to take advantage and create your own custom ones. All of this in simple, straightforward and understandable examples.

By the end of this workshop we will be able to do the following:

-   Use built-in React Hooks to create more reusable and straightforward code
-   Re-factor Class components into functional components using hooks
-   Create custom hooks and reuse them throughout a project
-   Know where React Hooks fall short and we should choose class components instead.

Speaker Bio
-----------

### João

João is a Frontend Engineer at Feedzai with over 4 years of developer experience and over 9 as an interface designer. Mostly, he just loves to design and build stuff for the Web. His designer/developer cross-breed skills are his main strength and help him to bring the best of both worlds to every project he gets involved into.

### Mickael

Mickael is Senior Frontend Engineer at Feedzai. He had the chance to work and experiment with a variety of languages and frameworks namely React. He was a teacher so he enjoys teaching and learning with a good conversation about technology. He has other tastes like sports (futebol and Krav Maga) and travel.

Links
-----

### João

-   Company: https://feedzai.com
-   GitHub: https://github.com/joaotmdias
-   Personal: https://joaodias.me
-   LinkedIn: https://www.linkedin.com/in/joaotmdias/

### Mickael

-   Company: https://feedzai.com
-   GitHub: https://github.com/Mickael24
-   LinkedIn: https://www.linkedin.com/in/dinomickael/

Extra Information
-----------------

### João

-   [Speaker at ENEMM 2019 with the keynote "State of the Web in 2019"](https://www.enemm.pt/speakers/joao-dias)
-   _Workshop: The User's Perception of Speed_ (2016)
-   _Keynote: State of Javascript 2018: Angular or Vue or React?_ (2018)
-   _Keynote: Let's talk about Accessibility_ (2019)

### Mickael

-   ["First React Application" workshop teacher at FEUP Code Week 2019"](https://www.linkedin.com/posts/dinomickael_feup-codeweek-feedzai-activity-6592090630468448256-k9Jk/)
-   Assistant Professor at the Polytechnic Institute of Cávado and Ave